chore: moves OptimismCard to its own file
